third_party/WebKit/LayoutTests/fast/dom/shadow/content-reprojection-recalc.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<div id="host">
    <div></div>
    <div></div>
    <div></div>
    <div></div>
    <div></div>
    <div></div>
</div>
<script>
description("Redistribution into same position should not cause style recalc");

shouldBeDefined("window.internals");

var root = host.createShadowRoot();
root.innerHTML = "<content></content>";

host.offsetTop;

host.appendChild(document.createElement("div"));

shouldBe("internals.updateStyleAndReturnAffectedElementCount()", "1");

</script>
