third_party/WebKit/LayoutTests/fast/dom/shadow/content-reprojection-recalc-expected.txt

Redistribution into same position should not cause style recalc

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.internals is defined.
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS successfullyParsed is true

TEST COMPLETE
